scst/include/backport.h: Suppress a compiler warning (merge r7881 from trunk)
git-svn-id: http://svn.code.sf.net/p/scst/svn/branches/3.3.x@7883 d57e44dd-8a1f-0410-8b47-8ef2f437770f

+++ scst/include/backport.h
@@ -893,7 +893,7 @@ struct t10_pi_tuple {
 #undef DEFINE_TIMER
 #define DEFINE_TIMER(_name, _function)					\
 	struct timer_list _name = TIMER_INITIALIZER(			\
-		(void (*)(unsigned long))(_function), 0,		\
+		(void (*)(unsigned long))(unsigned long)(_function), 0,	\
 		(unsigned long)&(_name))
 #endif
